Git (Gitolite) repositories were unexpectedly deleted when projects were updated
When a project is updated, the repository may also need to be updated. One of such cases arises when the project is moved to be subproject of another main project. The repository shall be moved in the filesystem to reflect the changes in the project.

The plugin did not handle this situation very well and in some cases the repository was deleted and replaced with a brand new EMPTY repository; this leaded to a complete loss of the original repository.